신고기능,관리자신고처리,신고알림
신고기능: 편지를 주고받을때 받은 사용자가 불쾌함을 느낀다면 신고할수 있다. 신고된 편지에 대한 정보는 관리자에게 전송
관리자 신고처리: 관리자는 신고된 편지 내용을 보고 판단 아래에 사용자를 신고 처리하면 사용자의 신고 누적 횟수가 추가된다.
(신고된 게시물은 중복 신고등을 위해 신고처리후 자동으로 삭제한다)
사용자는 일정 횟수 이상 신고를 받으면 로그인 후 경고 메시지를 받을 수 있다

+++ src/main/java/com/spring/tomorrow/mail/controller/MailControllerImpl.java
@@ -134,7 +134,7 @@ public class MailControllerImpl implements MailController {
 	}
 
 	@RequestMapping(value="/mail/sendReply.do", method=RequestMethod.POST)
-	public ResponseEntity sendReply(MailVO mailVO, HttpServletRequest request, HttpServletResponse response)
+	public ResponseEntity sendReply(@ModelAttribute MailVO mailVO, HttpServletRequest request, HttpServletResponse response)
 			throws Exception {
 		String message = null;
 		ResponseEntity resEntity = null;
@@ -192,14 +192,15 @@ public class MailControllerImpl implements MailController {
 		return mav;
 	}
 
-	@RequestMapping(value="/mail/deleteReceiveReply.do",method=RequestMethod.POST)
+	@RequestMapping(value= {"/mail/deleteReceiveReply.do","/mail/deleteReceiveReport"},method= {RequestMethod.POST,RequestMethod.GET})
 	public ResponseEntity deleteReceiveMail(@RequestParam int idx, HttpServletRequest request, HttpServletResponse response)
 			throws Exception {
 		String message = null;
 		ResponseEntity resEntity = null;
 		HttpHeaders responseHeaders = new HttpHeaders();
 		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
-		try {
+		if(request.getServletPath().equals("/mail/deleteReceiveReply.do")) {
+			try {
 			mailService.deleteReceiveMail(idx);
 			
 		    message  = "<script>";
@@ -207,12 +208,29 @@ public class MailControllerImpl implements MailController {
 		    message += " location.href='"+request.getContextPath()+"/mail/receiveReplyList.do';";
 		    message += " </script>";
 		    
-		}catch(Exception e) {
+			}catch(Exception e) {
 			message  = "<script>";
 		    message +=" alert('작업 중 오류가 발생했습니다. 다시 시도해 주세요');";
 		    message += " location.href='"+request.getContextPath()+"/mail/getReceiveReply.do?idx="+idx+"';";
 		    message += " </script>";
 			e.printStackTrace();
+			}
+		}
+		else {
+			try {
+				mailService.deleteReceiveMail(idx);
+				
+			    message  = "<script>";
+			    message += " location.href='"+request.getContextPath()+"/admin/reportList.do';";
+			    message += " </script>";
+			    
+				}catch(Exception e) {
+				message  = "<script>";
+			    message +=" alert('작업 중 오류가 발생했습니다. 다시 시도해 주세요');";
+			    message += " location.href='"+request.getContextPath()+"/admin/getReport.do?idx="+idx+"';";
+			    message += " </script>";
+				e.printStackTrace();
+				}
 		}
 		resEntity =new ResponseEntity(message, responseHeaders, HttpStatus.OK);
 		return resEntity;
@@ -262,6 +280,79 @@ public class MailControllerImpl implements MailController {
 		}
 		resEntity =new ResponseEntity(message, responseHeaders, HttpStatus.OK);
 		return resEntity;
+	}
+
+	@RequestMapping(value= {"/mail/reportWorry.do","/mail/reportReply.do"},method=RequestMethod.POST)
+	public ResponseEntity report(@ModelAttribute MailVO mailVO, HttpServletRequest request, HttpServletResponse response)
+			throws Exception {
+		String message = null;
+		ResponseEntity resEntity = null;
+		HttpHeaders responseHeaders = new HttpHeaders();
+		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+		if(request.getServletPath().equals("/mail/reportWorry.do")) {
+			try {
+				mailService.report(mailVO);
+			
+				message  = "<script>";
+				message +=" alert('해당 편지가 신고되었습니다');";
+				message += " location.href='"+request.getContextPath()+"/mail/receiveWorryList.do';";
+				message += " </script>";
+		    
+			}catch(Exception e) {
+				message  = "<script>";
+				message +=" alert('작업 중 오류가 발생했습니다. 다시 시도해 주세요');";
+				message += " location.href='"+request.getContextPath()+"/mail/getReceiveWorry.do?idx="+mailVO.getIdx()+"';";
+				message += " </script>";
+				e.printStackTrace();
+			}
+		
+		}
+		else {
+			try {
+				mailService.report(mailVO);
+			
+				message  = "<script>";
+				message +=" alert('해당 편지가 신고되었습니다');";
+				message += " location.href='"+request.getContextPath()+"/mail/receiveReplyList.do';";
+				message += " </script>";
+		    
+			}catch(Exception e) {
+				message  = "<script>";
+				message +=" alert('작업 중 오류가 발생했습니다. 다시 시도해 주세요');";
+				message += " location.href='"+request.getContextPath()+"/mail/getReceiveReply.do?idx="+mailVO.getIdx()+"';";
+				message += " </script>";
+				e.printStackTrace();
+			}
+		
+		}
+		resEntity =new ResponseEntity(message, responseHeaders, HttpStatus.OK);
+		return resEntity;
+	}
+
+	@RequestMapping(value="/admin/reportList",method=RequestMethod.GET)
+	public ModelAndView reportList(@RequestParam(defaultValue="1")int curPage, HttpServletRequest request, HttpServletResponse response)
+			throws DataAccessException {
+		HttpSession session = request.getSession();
+		MemberVO member=(MemberVO) session.getAttribute("member");
+		ModelAndView mav=new ModelAndView();
+		int totalCount=mailService.reportCount(member.getIdx());
+		Paging paging=new Paging(totalCount, curPage);
+		List<MailVO> reportList=mailService.reportList(member.getIdx(), paging.getStartIndex(),paging.getEndIndex());
+		mav.addObject("paging", paging);
+		mav.addObject("reportList", reportList);
+		mav.setViewName("/admin/reportList");
+		return mav;
+	}
+	
+	@RequestMapping(value="/admin/getReport.do",method=RequestMethod.GET)
+	public ModelAndView getReport(@RequestParam int idx, HttpServletRequest request, HttpServletResponse response)
+			throws DataAccessException {
+		ModelAndView mav=new ModelAndView();
+		MailVO report=mailService.getReceiveMail(idx);
+		mav.addObject("report", report);
+		mav.setViewName("/admin/getReport");
+		
+		return mav;
 	}
 
 }
